requirements.txt
azure-identity
openai
python-dotenv
lunr
requests
pymupdf4llm
langchain-text-splitters
tiktoken>=0.7.0
matplotlib
pandas
Pillow
networkx>=3.0
neo4j>=5.0.0